# 2024.12.03 Study ) 백준 13241번 문제

+++ 2024.12.03/Baekjoon_13241.java
@@ -5,10 +5,10 @@ public class Baekjoon_13241 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
-
-        long A = Integer.parseInt(st.nextToken());
-        long B = Integer.parseInt(st.nextToken());
-
+        
+        long A = Long.parseLong(st.nextToken());
+        long B = Long.parseLong(st.nextToken());
+        
         lcm(A, B);
     }
